SDL_GetBasePath, silence signed/unsigned compare
This warning should be fixed in SDL since we're just borrowing their
code wholesale for the moment, but rc is already known to be > 0, so
it's safe to simply cast it here.

+++ src/shim.cpp
@@ -144,7 +144,7 @@ SDL_GetBasePath(void)
             const int rc = (int) SDL_snprintf(path, sizeof(path),
                                               "/proc/%llu/exe",
                                               (unsigned long long) getpid());
-            if ( (rc > 0) && (rc < sizeof(path)) ) {
+            if ( (rc > 0) && ((unsigned)rc < sizeof(path)) ) {
                 retval = readSymLink(path);
             }
         }
